requirements.txt
numpy==1.19.5
jieba==0.42.1
matplotlib==3.3.4
Pillow==8.1.2
wordcloud==1.8.1 